+++ alibi_detect/saving/schemas.py
@@ -905,6 +905,7 @@ class LearnedKernelDriftConfig(DriftDetectorConfig):
     batch_size_predict: int = 1000000
     preprocess_batch_fn: Optional[str] = None
     epochs: int = 3
+    num_workers: int = 0
     verbose: int = 0
     train_kwargs: Optional[dict] = None
     dataset: Optional[str] = None
@@ -938,6 +939,7 @@ class LearnedKernelDriftConfigResolved(DriftDetectorConfigResolved):
     batch_size_predict: int = 1000000
     preprocess_batch_fn: Optional[Callable] = None
     epochs: int = 3
+    num_workers: int = 0
     verbose: int = 0
     train_kwargs: Optional[dict] = None
     dataset: Optional[Callable] = None
@@ -1263,7 +1265,7 @@ class RegressorUncertaintyDriftConfigResolved(DetectorConfig):
 
 
 # Unresolved schema dictionary (used in alibi_detect.utils.loading)
-DETECTOR_CONFIGS = {
+DETECTOR_CONFIGS: Dict[str, Type[DetectorConfig]] = {
     'KSDrift': KSDriftConfig,
     'ChiSquareDrift': ChiSquareDriftConfig,
     'TabularDrift': TabularDriftConfig,
@@ -1281,11 +1283,11 @@ DETECTOR_CONFIGS = {
     'FETDriftOnline': FETDriftOnlineConfig,
     'ClassifierUncertaintyDrift': ClassifierUncertaintyDriftConfig,
     'RegressorUncertaintyDrift': RegressorUncertaintyDriftConfig,
-}  # type: Dict[str, Type[DetectorConfig]]
+}
 
 
 # Resolved schema dictionary (used in alibi_detect.utils.loading)
-DETECTOR_CONFIGS_RESOLVED = {
+DETECTOR_CONFIGS_RESOLVED: Dict[str, Type[DetectorConfig]] = {
     'KSDrift': KSDriftConfigResolved,
     'ChiSquareDrift': ChiSquareDriftConfigResolved,
     'TabularDrift': TabularDriftConfigResolved,
@@ -1303,4 +1305,4 @@ DETECTOR_CONFIGS_RESOLVED = {
     'FETDriftOnline': FETDriftOnlineConfigResolved,
     'ClassifierUncertaintyDrift': ClassifierUncertaintyDriftConfigResolved,
     'RegressorUncertaintyDrift': RegressorUncertaintyDriftConfigResolved,
-}  # type: Dict[str, Type[DetectorConfig]]
+}