+++ alibi_detect/saving/tests/test_saving.py
@@ -486,7 +486,8 @@ def test_save_learnedkernel(data, deep_kernel, backend, tmp_path, seed):  # noqa
                                 deep_kernel,
                                 p_val=P_VAL,
                                 backend=backend,
-                                train_size=0.7)
+                                train_size=0.7,
+                                num_workers=0)
         preds = cd.predict(X_h0)  # noqa: F841
     save_detector(cd, tmp_path)
     with fixed_seed(seed):
@@ -806,6 +807,138 @@ def test_save_onlinefetdrift(data, tmp_path, seed):
     assert cd_load.n_bootstraps == N_BOOTSTRAPS
     assert cd_load.ert == ERT
     np.testing.assert_array_almost_equal(stats, stats_load, 4)
+
+
+@parametrize("detector", [MMDDriftOnline, LSDDDriftOnline])
+@parametrize_with_cases("data", cases=ContinuousData, prefix='data_')
+def test_save_multivariate_online_state(detector, data, backend, seed, tmp_path):
+    """
+    Test the saving (and loading) of multivariate online detectors' state via `save_detector`.
+    """
+    # Skip if backend not `tensorflow` or `pytorch`
+    if backend not in ('tensorflow', 'pytorch'):
+        pytest.skip("Detector doesn't have this backend")
+
+    # Init detector and make prediction to update state
+    X_ref, X_h0 = data
+    with fixed_seed(seed):
+        dd = detector(X_ref, ert=100, window_size=10, backend=backend)
+
+    # Run for 10 time-steps
+    test_stats = []
+    for t, x_t in enumerate(X_h0[:10]):
+        if t == 5:
+            # Save detector (with state)
+            save_detector(dd, tmp_path)
+        test_stats.append(dd.predict(x_t)['data']['test_stat'])
+
+    # Check state file created
+    assert dd._detector.state_dir == tmp_path.joinpath('state')
+
+    # Load
+    with fixed_seed(seed):
+        dd_new = load_detector(tmp_path)
+    # Check attributes and compare predictions at t=5
+    assert dd_new.t == 5
+    if detector == LSDDDriftOnline:  # Often a small (~1e-6) difference in LSDD test stats post-load # TODO - why?
+        np.testing.assert_array_almost_equal(dd_new.predict(X_h0[5])['data']['test_stat'], test_stats[5], 5)
+    else:
+        np.testing.assert_array_equal(dd_new.predict(X_h0[5])['data']['test_stat'], test_stats[5])
+
+    # Check that error raised if no state file inside `state/` dir
+    for child in tmp_path.joinpath('state').glob('*'):
+        if child.is_file():
+            child.unlink()
+    with pytest.raises(FileNotFoundError):
+        load_detector(tmp_path)
+
+
+@parametrize("detector", [CVMDriftOnline])
+@parametrize_with_cases("data", cases=ContinuousData, prefix='data_')
+def test_save_cvm_online_state(detector, data, tmp_path):
+    """
+    Test the saving (and loading) of the CVM online detector's state via `save_detector`.
+    """
+    # Init detector and make prediction to update state
+    X_ref, X_h0 = data
+    dd = detector(X_ref, ert=100, window_sizes=[10])
+
+    # Run for 10 time-steps
+    test_stats = []
+    for t, x_t in enumerate(X_h0[:10]):
+        if t == 5:
+            # Save detector (with state)
+            save_detector(dd, tmp_path)
+        test_stats.append(dd.predict(x_t)['data']['test_stat'])
+
+    # Check state file created
+    assert dd.state_dir == tmp_path.joinpath('state')
+
+    # Load
+    dd_new = load_detector(tmp_path)
+    # Check attributes and compare predictions at t=5
+    assert dd_new.t == 5
+    np.testing.assert_array_equal(dd_new.predict(X_h0[5])['data']['test_stat'], test_stats[5])
+
+    # Check that error raised if no state file inside `state/` dir
+    for child in tmp_path.joinpath('state').glob('*'):
+        if child.is_file():
+            child.unlink()
+    with pytest.raises(FileNotFoundError):
+        load_detector(tmp_path)
+
+
+@parametrize("detector", [FETDriftOnline])
+@parametrize_with_cases("data", cases=BinData, prefix='data_')
+def test_save_fet_online_state(detector, data, tmp_path):
+    """
+    Test the saving (and loading) of the FET online detector's state via `save_detector`.
+    """
+    # Init detector and make prediction to update state
+    X_ref, X_h0 = data
+    dd = detector(X_ref, ert=100, window_sizes=[10])
+
+    # Run for 10 time-steps
+    test_stats = []
+    for t, x_t in enumerate(X_h0[:10]):
+        if t == 5:
+            # Save detector (with state)
+            save_detector(dd, tmp_path)
+        test_stats.append(dd.predict(x_t)['data']['test_stat'])
+
+    # Check state file created
+    assert dd.state_dir == tmp_path.joinpath('state')
+
+    # Load
+    dd_new = load_detector(tmp_path)
+    # Check attributes and compare predictions at t=5
+    assert dd_new.t == 5
+    np.testing.assert_array_equal(dd_new.predict(X_h0[5])['data']['test_stat'], test_stats[5])
+
+    # Check that error raised if no state file inside `state/` dir
+    for child in tmp_path.joinpath('state').glob('*'):
+        if child.is_file():
+            child.unlink()
+    with pytest.raises(FileNotFoundError):
+        load_detector(tmp_path)
+
+
+@parametrize_with_cases("data", cases=ContinuousData, prefix='data_')
+def test_save_online_state_t0(data, tmp_path):
+    """
+    Test that state is not saved when t=0.
+    """
+    # Init detector
+    X_ref, X_h0 = data
+    dd = CVMDriftOnline(X_ref, ert=100, window_sizes=[10])
+    # Check state NOT saved when t=0
+    state_dir = tmp_path.joinpath('state')
+    save_detector(dd, tmp_path)
+    assert not state_dir.is_dir()
+    # Check state IS saved when t>0
+    dd.predict(X_h0[0])
+    save_detector(dd, tmp_path)
+    assert state_dir.is_dir()
 
 
 @parametrize_with_cases("data", cases=ContinuousData.data_synthetic_nd)